# IGNITE-20745 Improve performance of ClientTableCommon.readTableAsync