--- a/modules/client/src/test/java/org/apache/ignite/client/fakes/FakeIgniteTables.java
+++ b/modules/client/src/test/java/org/apache/ignite/client/fakes/FakeIgniteTables.java
@@ -39,6 +39,7 @@ import org.apache.ignite.internal.tx.impl.HeapLockManager;
 import org.apache.ignite.internal.type.NativeTypes;
 import org.apache.ignite.lang.IgniteException;
 import org.apache.ignite.table.Table;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * Fake tables.
@@ -160,6 +161,11 @@ public class FakeIgniteTables implements IgniteTablesInternal {
     @Override
     public CompletableFuture<TableViewInternal> tableViewAsync(String name) {
         return completedFuture(tableView(name));
+    }
+
+    @Override
+    public @Nullable TableViewInternal cachedTable(int tableId) {
+        return table(tableId);
     }
 
     private TableViewInternal getNewTable(String name, int id) {