--- a/modules/platforms/dotnet/Apache.Ignite.Benchmarks/Program.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Benchmarks/Program.cs
@@ -22,5 +22,6 @@ using Table;
 
 internal static class Program
 {
-    private static void Main() => BenchmarkRunner.Run<DataStreamerBenchmark>();
+    // IMPORTANT: Disable Netty leak detector when using a real Ignite server for benchmarks.
+    private static void Main() => BenchmarkRunner.Run<TupleGetBenchmarks>();
 }