* `ClientTableCommon.readTableAsync` is called on every table operation (get, put, etc), so performance is critical
* It calls `TableManager.tableAsync(int tableId)`, which performs complex async logic
* To speed this up, check cached tables in `TableManager.startedTables`, which will return a table immediately in most cases
  * The table can be stopped, but that is not a problem:
    * New table with the same id can't be started
    * Any operation will throw an exception

Benchmarked from .NET thin client (`TupleGetBenchmarks`), ~8% improvement:

```
| Method |     Mean |   Error |  StdDev |
|------- |---------:|--------:|--------:|
|    Get | 123.6 us | 2.46 us | 5.66 us | (before)
|    Get | 113.8 us | 1.74 us | 1.45 us | (after)
```

https://issues.apache.org/jira/browse/IGNITE-20745